+++ src/test/ExSln2/c99/gen/hal_Led.h
@@ -3,7 +3,6 @@
 
 #include "hal_Gpio.h"
 #include <stdint.h>
-#include "hal_Led.h"