+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Test/interim releases are not documented here.
 ### Fixed
 - Transpiler - now tracks header dependency of type `T` contained in `c_array<T>`
 - Transpiler - fix pass `self` to method calls like `this.toggle()`
+- Transpiler - prevent .h file from including itself in the same file
 
 ---